code review suggestions

+++ contracts/cheddar_5x5_tic_tac_toe/src/utils.rs
@@ -19,8 +19,8 @@ pub(crate) type GameId = u64;
 pub(crate) type AffiliateId = AccountId;
 
 /// This constant can be used to set the board size
-pub(crate) const BOARD_SIZE: usize = 25;
-pub(crate) const MAX_NUM_TURNS: u64 = (BOARD_SIZE*BOARD_SIZE) as u64;
+pub(crate) const BOARD_SIZE: u8 = 25;
+pub(crate) const MAX_NUM_TURNS: u64 = BOARD_SIZE as u64 *BOARD_SIZE as u64;
 
 
 /// Returns true if the promise was failed. Otherwise returns false.
